# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Project Final",
    "version" : "v1",
    "title" : "ProjectFinal",
    "contact" : {
      "name" : "tesis",
      "email" : "leandro.cocchi@davinci.com"
    }
  },
  "host" : "*.*",
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/rol/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/user" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/version/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Version"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Version"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Version"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Rol" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
...
    "User" : {
      "type" : "object",
      "required" : [ "created", "enabled", "firstName", "lastName", "password", "rol", "updated", "userName" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "firstName" : {
          "type" : "string",
          "minLength" : 1
        },
        "lastName" : {
          "type" : "string",
          "minLength" : 1
        },
        "dni" : {
          "type" : "integer",
          "format" : "int32"
        },
        "email" : {
          "type" : "string"
        },
        "userName" : {
          "type" : "string",
          "minLength" : 1
        },
        "rol" : {
          "type" : "integer",
          "format" : "int32"
        },
        "password" : {
          "type" : "string",
          "minLength" : 1
        },
        "avatar" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updated" : {
          "type" : "string",
          "format" : "date-time"
        },
        "enabled" : {
          "type" : "boolean"
        }
      }
    },
    "Version" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    }
  }
}